+++ src/pages/Dashboard/AlertRadio.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { AlertComponent } from '@components/layout/AlertComponent';
+import { AlertComponent } from '@components/layout/AlertComponent/AlertComponent';
 
 export const AlertRadio: React.FC = () => {
   const handleOk = (data: { [key: string]: any }): void => {
