+++ src/store/AlertStore/IAlertButton.ts
@@ -1,0 +1,8 @@
+/**
+ * Interface for alert button properties.
+ */
+export interface IAlertButton {
+  text: string;
+  role?: 'cancel' | 'destructive';
+  handler?: (value?: any) => void;
+}
