+++ src/store/AlertStore/IAlertInput.ts
@@ -1,0 +1,7 @@
+import { ITextInput } from './ITextInput';
+import { IRadioInput } from './IRadioInput';
+
+/**
+ * Union type for alert inputs.
+ */
+export type IAlertInput = ITextInput | IRadioInput;
