+++ src/store/AlertStore/IAlertOptions.ts
@@ -1,0 +1,11 @@
+import { IAlertInput } from './IAlertInput';
+import { IAlertButton } from './IAlertButton';
+
+/**
+ * Interface for alert options.
+ */
+export interface IAlertOptions {
+  subHeader?: string;
+  buttons?: (string | IAlertButton)[];
+  inputs?: IAlertInput[];
+}
