+++ src/store/AlertStore/IAlertState.ts
@@ -1,0 +1,13 @@
+import { IAlertOptions } from './IAlertOptions';
+
+/**
+ * Interface for alert state.
+ */
+export interface IAlertState {
+  showAlert: (
+    header: string,
+    message: string,
+    options?: IAlertOptions,
+  ) => Promise<void>;
+  closeAlert: () => Promise<void>;
+}
