+++ src/store/AlertStore/IFormData.ts
@@ -1,0 +1,6 @@
+export interface IFormData {
+  name: string;
+  nickname: string;
+  age: number;
+  about: string;
+}
